Change hash_t from uint64_t to uint32_t

+++ src/hash/hash.h
@@ -3,7 +3,7 @@
 
 #include <stdint.h>
 
-typedef uint64_t hash_t;
+typedef uint32_t hash_t;
 
 /**
  *
